Updates the Ruby specs to operate again
The lib/main.rb was doing two things. Acting as the central switchboard
to load all the files for the project and then executing the gem. To get
the tests to work there needed to be the first thing but not the second
otherwise the game would start and then you would have to quit it for the
tests to execute. So the 'lib/adventure.rb' is born which loads all the
ruby files and now 'lib/main.rb' will require that file and start the game.

To test this initial arrangement I wrote a single test file for the
Location Struct.

I removed the previous test related to the map specification because it
referenced a few constants that no longer exist. When creating the tests
for the Room and such this file can be revisited in the history to see
if those tests can be resurrected and rebuilt.